refactor(style): DocumentPaint graduates from document.chart to document.style
The paint vocabulary (sealed Solid | Linear | Radial with colour stops) is
not chart-specific - it is the shared fill type every fillable surface will
use. Pre-release move (the v1.8.0 cycle is unreleased), so no deprecation
bridge: chart classes, tests, examples re-import from document.style. The
javadoc drops the 'lives in the chart package for now' note and states the
backend contract instead: gradient-capable backends paint shadings, others
degrade to primaryColor(). Zero behaviour change - full suite green
unchanged.
